Change mgo package for juju compatibility.
Juju has changed the import path of the mgo package to the juju fork,
since the original is no longer maintained. This updates mgosession to
support the same change. As this changes the return types of the package
the module version has been incremented to maintain compatibility for
any old builds.

Also replace the tomb dependency with a context and waitgroup to reduce
the number of external dependencies.